mv examples out of docs repo
