+++ apps/docs/examples/inputs.mdx
