+++ apps/docs/examples/switch-slider-rating.mdx
