+++ apps/rhf-mui-docs/docs/introduction.mdx
@@ -42,7 +42,7 @@ to install [@mui/x-date-pickers](https://www.npmjs.com/package/@mui/x-date-picke
 
 ## Working
 Each component requires a set of common props, detailed in the table below.
-Some props might be specific to a component, say [DatePicker](../components/pickers/DatePicker#api), 
+Some props might be specific to a component, say [DatePicker](../components/pickers/RHFDatePicker#api), 
 and can be viewed in the documentation page for the respective component.
 
 _You can also import the interface of each component, like `RHFDatePickerProps`
