+++ apps/rhf-mui-docs/src/pages/index.tsx
@@ -40,9 +40,9 @@ function HomepageHeader() {
               justifyContent: 'flex-start'
             }}
           >
-            <Link className="button button--lg" to="/intro">
+            {/* <Link className="button button--lg" to="/intro"> */}
               View Demo
-            </Link>
+            {/* </Link> */}
           </Grid>
           <Button color='secondary' variant='contained'>hell3</Button>
         </Grid>
